+++ src/slash-commands/help.ts
@@ -1,0 +1,35 @@
+/*
+Copyright 2015, 2016 OpenMarket Ltd
+Copyright 2018 New Vector Ltd
+Copyright 2019 Michael Telatynski <7t3chguy@gmail.com>
+Copyright 2020 The Matrix.org Foundation C.I.C.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import { _td } from "../languageHandler";
+import Modal from "../Modal";
+import SlashCommandHelpDialog from "../components/views/dialogs/SlashCommandHelpDialog";
+import { success } from "./utils";
+import { Command } from "./command";
+import { CommandCategories } from "./interface";
+
+export const help = new Command({
+    command: "help",
+    description: _td("Displays list of commands with usages and descriptions"),
+    runFn: function () {
+        Modal.createDialog(SlashCommandHelpDialog);
+        return success();
+    },
+    category: CommandCategories.advanced,
+});
