+++ src/slash-commands/rageshake.ts
@@ -1,0 +1,42 @@
+/*
+Copyright 2015, 2016 OpenMarket Ltd
+Copyright 2018 New Vector Ltd
+Copyright 2019 Michael Telatynski <7t3chguy@gmail.com>
+Copyright 2020 The Matrix.org Foundation C.I.C.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+import { _td } from "../languageHandler";
+import SdkConfig from "../SdkConfig";
+import { success } from "./utils";
+import Modal from "../Modal";
+import BugReportDialog from "../components/views/dialogs/BugReportDialog";
+import { Command } from "./command";
+import { CommandCategories } from "./interface";
+
+export const rageshake = new Command({
+    command: "rageshake",
+    aliases: ["bugreport"],
+    description: _td("Send a bug report with logs"),
+    isEnabled: () => !!SdkConfig.get().bug_report_endpoint_url,
+    args: "<description>",
+    runFn: function (cli, roomId, threadId, args) {
+        return success(
+            Modal.createDialog(BugReportDialog, {
+                initialText: args,
+            }).finished,
+        );
+    },
+    category: CommandCategories.advanced,
+});
